deleting old files
Since we can still look at older commits
