+++ code/main.py
@@ -1,83 +1,102 @@
-from shapes import Triangle, Shape
-from hexshapes import Hexagon, HShape
-from shapes import hexagon_maker as hexmaker
-from shapes import triangle_of_hexes as toh
-from hexshapes import bighex_maker as bhmaker
+from squareshapes import Square, Polyomino, bigsquare_maker
+import matplotlib.pyplot as plt
+from matplotlib.patches import RegularPolygon
+import numpy as np
+import ast
 import time
 
-def data_writer(base, type = "corona"):
-    with open('./code/plotlist.txt', 'w') as file:
-        if type == "heesch":
-            coronalist = base.heesch_computer()[0]
-            coronadata = []
-            for corona_config in coronalist:
-                coronadata.append([shape.to_data() for shape in corona_config])
-            datadict = {"type": type, "num": len(coronalist), "data": {"base": base.to_data(), "heesch": coronadata}}
-
-        elif type == "corona2":
-            coronalist = base.second_corona()[0]
-            first_data = []
-            for shape in coronalist["first"]:
-                first_data.append(shape.to_data())
-            second_data = []
-            for shape in coronalist["second"][0]:
-                second_data.append(shape.to_data())
-            datadict = {"type": type, "data": {"base": base.to_data(), "first": first_data, "second": second_data}}
-
-        elif type == "corona":
-            coronalist = base.corona_maker(base.orientations())[0]
-            corona_data = []
-            for shape in coronalist:
-                corona_data.append(shape.to_data())
-            datadict = {"type": type, "data": {"base": base.to_data(), "corona": corona_data}}
-
-        elif type == "base":
-            datadict = {"type": type, "data": {"base": base.to_data()}}
-
-        file.write(str(datadict))
-
-start_time = time.time()
-
-""" Some testing shapes"""
-# S1 = Shape(toh())
-# S1 = Shape(toh() + hexmaker(2,-5) + hexmaker(5, -2) + hexmaker(6, -3))
-# S1 = Shape([triangle for triangle in toh() if triangle not in hexmaker(3,-3)+hexmaker(2,-2)])
-# S1 = Shape([triangle for triangle in hexmaker(0,0)+[Triangle(1,1)] if triangle not in [Triangle(0,0)]])
-# S1 = Shape(hexmaker(-1,1)+hexmaker(1,2)+hexmaker(2,1)+hexmaker(1,-1)+hexmaker(-1,-2))
 
 
-""" edge data testing """
-# S1 = HShape([Hexagon(0,0, [1, 1, 1, 0, -1, -1])])
+def main_func(tile):
 
-S1 = HShape([
-Hexagon(0, 0, [0, 0, 0, 1, -1, 0]),
-Hexagon(1, -1),
-Hexagon(2, 1),
-Hexagon(3, 0),
-])
+    def plotter(plottinglist, patches):
+        axs = plt.subplot()
+        for patch in patches:
+            axs.add_patch(patch)
+        axs.plot(*(plottinglist), linewidth=0.5)
+        plt.autoscale(enable = True)
+        axs.set_aspect("equal")
+        plt.axis('off')
+        plt.show()
 
-# H1 = Hexagon(0, 0, [1, 0, 0, 0, 0, 0])
-# H2 = Hexagon(1, 2, [0, 0, -1, 0, 0, 0])
-# S1 = HShape([H1, H2])
+    def has_holes(config, output = False ):
+        total_squares = set()
+        for shape in config+[tile]:
+            total_squares = total_squares.union({(square.origin[0],square.origin[1]) for square in shape.squares})
+        extended_squares = set()
+        for square in total_squares:
+            extended_squares = extended_squares.union({
+            square,
+            (square[0], square[1]-2),
+            (square[0]+2, square[1]-2),
+            (square[0]+2, square[1]),
+            (square[0]+2, square[1]+2),
+            (square[0], square[1]+2),
+            (square[0]-2, square[1]+2),
+            (square[0]-2, square[1]),
+            (square[0]-2, square[1]-2),
+            })
+        without_inside = extended_squares.difference(total_squares)
 
-# S1 = HShape([
-# Hexagon(0, 0, [0, 0, 0, 0, -1, 0]),
-# Hexagon(1, -1, [0, 0, 0, 0, 0, 1]),
-# Hexagon(2, 1),
-# Hexagon(3, 0),
-# Hexagon(0, -3),
-# ])
+        # getting a starting position
+        xmin = min({coord[0] for coord in without_inside})
+        ymin = min({coord[1] for coord in without_inside if coord[0] == xmin})
+        mincoord = (xmin, ymin)
 
-""" hexagon testing """
-# S1 = HShape( [Hexagon(0,0), Hexagon(2,1), Hexagon(3,0)])
-# S1 = HShape( [hex for hex in bhmaker(0,0) if hex not in [Hexagon(-2,-1), Hexagon(0,0)] ]  )
-# S1 = HShape( bhmaker(0,0)+[Hexagon(-2,2), Hexagon(-2,-4), Hexagon(4,2), Hexagon(0,-3), Hexagon(3,0), Hexagon(4,-1)] )
+        # computing the connected component and removing them
+        coords = { mincoord }
+        while coords != set():
+            without_inside = without_inside.difference(coords)
+            new_coords = set()
+            for coord in coords:
+                nextcoords = {
+                (coord[0], coord[1]-2),
+                (coord[0]+2, coord[1]),
+                (coord[0], coord[1]+2),
+                (coord[0]-2, coord[1]),
+                }
+                nextcoords = {n_coord for n_coord in nextcoords if n_coord in without_inside}
+                new_coords = new_coords.union(nextcoords)
+            coords = new_coords
+
+        return without_inside if output else (without_inside != set())
+
+    def tileplotter(color, tile_=tile):
+        plottinglist.extend(tile_.plot_data())
+        for square in tile_.squares:
+            plot_square = RegularPolygon(square.origin, numVertices=4 ,orientation = 1/4 *np.pi, radius= np.sqrt(2), alpha=1, color= color)
+            patches.append(plot_square)
+
+    def c_config_plotter(c_config):
+
+        colordict = {
+        0: "mediumaquamarine",
+        1: "lightseagreen",
+        2: "teal",
+        3: "red",
+        4: "yellow",
+        }
+
+        tileplotter("aquamarine")
+
+        for c_config_num  in range(len(c_config)):
+            corona = c_config[c_config_num].copy()
+            for tile_ in corona:
+                tileplotter(color = colordict[c_config_num], tile_= tile_)
+
+    start_time = time.time()
+
+    plottinglist = []
+    patches = []
+
+    heesch_configs = tile.heesch_computer()
 
 
-""" This has types:
-    1) Corona
-    2) Boundary
-    3) Base
-"""
-data_writer(S1, type = "corona")
-print("--- %s seconds ---" % (time.time() - start_time))
+    c_config = heesch_configs[0]
+
+
+    c_config_plotter(c_config)
+
+
+    print("--- %s seconds ---" % (time.time() - start_time))
+    plotter(plottinglist, patches)
