--- code/plotlist.txt
@@ -1,1 +1,0 @@
-{'type': 'corona', 'data': {'base': [[0, 0, [0, 0, 0, 1, -1, 0]], [1, -1, [0, 0, 0, 0, 0, 0]], [2, 1, [0, 0, 0, 0, 0, 0]], [3, 0, [0, 0, 0, 0, 0, 0]]], 'corona': [[[-4, -5, [0, 0, 0, 0, 1, -1]], [-2, -4, [0, 0, 0, 0, 0, 0]], [-3, -3, [0, 0, 0, 0, 0, 0]], [-1, -2, [0, 0, 0, 0, 0, 0]]], [[1, 2, [-1, 0, 0, 0, 0, 1]], [2, 4, [0, 0, 0, 0, 0, 0]], [0, 3, [0, 0, 0, 0, 0, 0]], [1, 5, [0, 0, 0, 0, 0, 0]]], [[-1, 1, [0, 1, -1, 0, 0, 0]], [-3, 0, [0, 0, 0, 0, 0, 0]], [-2, -1, [0, 0, 0, 0, 0, 0]], [-4, -2, [0, 0, 0, 0, 0, 0]]], [[0, -6, [-1, 0, 0, 0, 0, 1]], [1, -4, [0, 0, 0, 0, 0, 0]], [-1, -5, [0, 0, 0, 0, 0, 0]], [0, -3, [0, 0, 0, 0, 0, 0]]], [[5, -2, [1, -1, 0, 0, 0, 0]], [4, -1, [0, 0, 0, 0, 0, 0]], [3, -3, [0, 0, 0, 0, 0, 0]], [2, -2, [0, 0, 0, 0, 0, 0]]], [[7, 2, [1, -1, 0, 0, 0, 0]], [6, 3, [0, 0, 0, 0, 0, 0]], [5, 1, [0, 0, 0, 0, 0, 0]], [4, 2, [0, 0, 0, 0, 0, 0]]], [[6, 6, [0, 1, -1, 0, 0, 0]], [4, 5, [0, 0, 0, 0, 0, 0]], [5, 4, [0, 0, 0, 0, 0, 0]], [3, 3, [0, 0, 0, 0, 0, 0]]]]}}
